move grouped data calculation to helper
