+++ workers/loc.api/sync/get-risk/helpers/calc-grouped-data.js
@@ -1,0 +1,102 @@
+'use strict'
+
+const { pick, omit, orderBy } = require('lodash')
+
+const _getDataKeys = (data) => {
+  return Object.keys(data)
+    .filter(key => (
+      Array.isArray(data[key]) &&
+      data[key].length > 0
+    ))
+}
+
+const _getMaxLength = (data) => {
+  const dataArr = Object.values(data)
+
+  if (dataArr.length === 0) {
+    return 0
+  }
+
+  return Math.max(
+    ...dataArr.map(item => item.length)
+  )
+}
+
+const _mergeData = (data) => {
+  if (
+    !data ||
+    typeof data !== 'object'
+  ) {
+    return []
+  }
+
+  const dataKeys = _getDataKeys(data)
+  const _data = pick(data, dataKeys)
+  const maxLength = _getMaxLength(_data)
+  const res = []
+
+  for (let i = 0; maxLength > i; i += 1) {
+    dataKeys.forEach(key => {
+      const { mts, vals } = { ..._data[key][i] }
+
+      if (
+        !Number.isInteger(mts) ||
+        !vals ||
+        typeof vals !== 'object' ||
+        Object.keys(vals).length === 0
+      ) {
+        return
+      }
+      if (
+        res.length === 0 ||
+        res.every(item => mts !== item.mts)
+      ) {
+        res.push({
+          mts,
+          [key]: { ...vals }
+        })
+
+        return
+      }
+
+      res.forEach((item, index) => {
+        if (mts === item.mts) {
+          res[index] = {
+            ...item,
+            [key]: { ...vals }
+          }
+        }
+      })
+    })
+  }
+
+  return orderBy(res, ['mts'], ['desc'])
+}
+
+module.exports = (data, isSubCalc) => {
+  const _data = _mergeData(data)
+  return _data.map(item => {
+    const res = Object.values(omit(item, ['mts']))
+      .reduce((accum, curr) => {
+        Object.entries(curr).forEach(([symb, val]) => {
+          if (!Number.isFinite(val)) {
+            return
+          }
+          if (Number.isFinite(accum[symb])) {
+            accum[symb] += val
+
+            return
+          }
+
+          accum[symb] = val
+        })
+
+        return accum
+      }, {})
+
+    return {
+      mts: item.mts,
+      ...(isSubCalc ? { vals: { ...res } } : res)
+    }
+  })
+}
