+++ workers/loc.api/sync/get-risk/helpers/index.js
@@ -1,7 +1,9 @@
 'use strict'
 
 const groupByTimeframe = require('./group-by-timeframe')
+const calcGroupedData = require('./calc-grouped-data')
 
 module.exports = {
-  groupByTimeframe
+  groupByTimeframe,
+  calcGroupedData
 }
